-Fixed bug that prevented loading of TYI,VGI, etc (sorry!) -Fixed bugs in automatic channel selection for polyphonic patches -Fixed bug where drumless VGMs would still import a drum track that would crash the plugin -Fixed text alignment for item names in presets/instrument list -Fixed LFO functionality (global LFO enable/amplitude is now a global patch-independent setting as it should be) -Swapped operators 2 and 3, as they where backwards (Thanks Plygon!)
V1.13 Changelog:
-Fixed bug where VST version could not log VGMs
-Fixed VGM DAC export errors
-Fixed VGM export tuning
-Drum instruments now remember their sample beginnings/endings and whether or not they loop
-Fixed drum note labels
-Fixed TRUE STEREO in VST
-Fixed VST pitch bending to work on all 16 channels
-VGM Logging now uses the lowest 2 notes instead of the highest 2.
-Removed import/export support for old DAC formats, since they wouldn't work anyway. Drums are now imported and exported using normal GennyInstrument.gen export
-All imported drums are now down sampled to 11025hz, since that's what GENNY was playing them at anyway. If you don't want to lose quality on import, convert your samples to 11025hz before importing.

V1.12 Changelog:
-Some optimizations
-Fixed pitch bending automation in FL

V1.11 Changelog:
-Improved channel management to reduce accidental note cut-offs when working within channel limits
-Fixed serious issue with instruments attack envelopes 'clicking' at high velocities
-Removed stray pixel on Alg 8
-Fixed bugs with copy-pasting patches, you can now copy paste drum patches
-Velocity calculation is now more accurate, this and the clicking fix unfortunately means existing project velocity values may need adjustment.
-Fixed display bugs with instrument list channel indicators

V 1.1 Changelog:
-Fixed some automation related crashes
-Fixed DAC panning
-Added MIDI pitch bend functionality in VST version
-New MEGA MIDI hardware mode, allowing the synth to interface with Aidan Lawrence' MEGA MIDI hardware!
-New CLEAN and ACCURATE emulation modes. ACCURATE is, well, more accurate but it's also more noisy.
-VGM Import! Import Bank can now import VGM/VGZ files and pull presets out of them, including limited support for drum samples.

+++ Genny/src/YM2612/WaveData.h
@@ -16,7 +16,7 @@ struct WaveData
 
 	int size;
 	unsigned char* audioData;
-	int audioPosition;
+	float audioPosition;
 	float waitTime;
 
 	int streamStartPos;
